first attempt at creating routes for the backend
