--- server/index.js
@@ -1,3 +1,0 @@
-const server = require("./server.js");
-
-server.listen(3000);
